nmad:   -- fix warnings in examples
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@27515 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/examples/pack/nm_pack_ping.c
@@ -87,8 +87,8 @@ int main(int argc, char**argv)
 	  bw_million_byte = len * (LOOPS / (sum / 2));
 	  bw_mbyte        = bw_million_byte / 1.048576;
 	  
-	  printf("%d\t%lf\t%8.3f\t%8.3f\n",
-		 len, lat, bw_million_byte, bw_mbyte);
+	  printf("%lld\t%lf\t%8.3f\t%8.3f\n",
+		 (long long)len, lat, bw_million_byte, bw_mbyte);
 	}
     }
